// ui/src/generated/PluginCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DeviceId } from "./DeviceId";
import type { PluginFormat } from "./PluginFormat";

export type PluginCommand = { "type": "Rescan", full?: boolean, } | { "type": "List" } | { "type": "OpenEditor", device: DeviceId, } | { "type": "CloseEditor", device: DeviceId, } | { "type": "SetSandboxed", device: DeviceId, sandboxed: boolean, } | { "type": "Reload", device: DeviceId, } | { "type": "ListFolders" } | { "type": "AddFolder", path: string, format: PluginFormat | null, } | { "type": "RemoveFolder", path: string, } | { "type": "SetIncludeDefaults", include: boolean, };